Started refactory of the code in order to build a clean and efficient tree function, or at least a blazing fast divide
