+++ src/test/integration/divide.rs
@@ -1,4 +1,6 @@
-use crate::{board::Board, debug::divide::divide, legal_moves::misc::Color};
+use crate::board::board::Board;
+use crate::board::fen_handling::FenHandling;
+use crate::{debug::divide::divide, legal_moves::misc::Color};
 
 #[test]
 fn perft() {
