+++ src/test/integration/test_games.rs
@@ -1,4 +1,4 @@
-use crate::board::Board;
+use crate::board::board::Board;
 use crate::legal_moves::is_move_possible::*;
 use crate::legal_moves::misc::{Color, Move};
 use crate::utils::move_to_string;
